migration type o'zgartirildi.

+++ migrations/4_2_add-balans.js
@@ -4,7 +4,7 @@
  */
 exports.up = function (knex) {
   return knex.schema.alterTable("driver", (table) => {
-    table.bigInteger("balans").defaultTo(10000);
+    table.float("balans").defaultTo(10000);
   });
 };
